+++ app/Http/Controllers/CategoryController.php
@@ -8,27 +8,24 @@ use App\Http\Requests\Category\UpdateRequest;
 use App\Models\Category;
 use App\Services\AttributeService;
 use App\Services\CategoryService;
-use Illuminate\Support\Facades\Config;
-use Illuminate\Support\Facades\DB;
-use Illuminate\Support\Facades\Storage;
 
 class CategoryController extends Controller
 {
-    public function __construct(public CategoryService $categoryService)
+    public function __construct(private CategoryService $categoryService)
     {
     }
 
     public function index()
     {
-        $categories = $this->categoryService->getWithLevels();
-        return view('category.index', compact('categories'));
+        return view('category.index', ['categories' => $this->categoryService->getWithLevels()]);
     }
 
     public function create()
     {
-        $categories = $this->categoryService->getWithLevels();
-        $imgParams = $this->categoryService->getImgParams();
-        return view('category.create', compact('categories', 'imgParams'));
+        return view('category.create', [
+            'categories' => $this->categoryService->getWithLevels(),
+            'imgParams' => $this->categoryService::getImgParams(),
+        ]);
     }
 
     public function store(StoreRequest $request)
@@ -38,41 +35,39 @@ class CategoryController extends Controller
             ->with('status', 'Категория "' . $category->title . '" успешно добавлена.');
     }
 
-    public function show(Category $category, AttributeService $attributeService)
+    public function edit(Category $category, AttributeService $attributeService)
     {
-        $categoryAttributes = $category->attributesWithUnitTitle();
-        $parentAttributesIds = $category->attributesIdsOfTheParentCategory();
-        $attributes = $attributeService->getWithUnitTitle();
-        return view('category.show', compact('category',
-            'categoryAttributes', 'parentAttributesIds', 'attributes'));
-    }
-
-    public function edit(Category $category)
-    {
-        $categories = $this->categoryService->getWithLevels();
-        $imgParams = $this->categoryService->getImgParams();
-        $parentCategoryId = $category->parentCategoryId();
-        $previousCategoryId = $category->previousCategoryId();
-        return view('category.edit', compact('category', 'categories',
-            'imgParams', 'parentCategoryId', 'previousCategoryId'));
+        return view('category.edit', [
+            'category' => $category,
+            'categories' => $this->categoryService->getWithLevels(),
+            'imgParams' => $this->categoryService::getImgParams(),
+            'parentCategoryId' => $category->parentCategoryId(),
+            'previousCategoryId' => $category->previousCategoryId(),
+            'attributes' => $attributeService->getWithUnitTitle(),
+            'categoryAttributesIds' => $category->attributesIds(),
+            'parentAttributesIds' => $category->attributesIdsOfTheParentCategory(),
+        ]);
     }
 
     public function update(UpdateRequest $request, Category $category)
     {
         $this->categoryService->change($category, $request->validated());
-        return redirect()->route('categories.show', $category->id)
+        return redirect()->route('categories.edit', $category->id)
             ->with('status', 'Категория успешно обновлена.');
     }
 
     public function updateAttributes(UpdateAttributesRequest $request, Category $category)
     {
-        $this->categoryService->changeAttributes($category, $request->validated());
-        return redirect()->route('categories.show', $category->id)
+        $this->categoryService->changeAttributes($category, $request->validated('attributes_ids'));
+        return redirect()->route('categories.edit', $category->id)
             ->with('status', 'Атрибуты категории обновлены.');
     }
 
     public function destroy(Category $category)
     {
+        if ($category->hasProductsOrAChildCategoryHasProducts()) {
+            return back()->withErrors(['categoryDeletion' => 'You cannot delete a category "' . $category->title . '" that is used by at least one product']);
+        }
         $this->categoryService->remove($category);
         return redirect()->route('categories.index')
             ->with('status', 'Категория "' . $category->title . '" удалена.');
